Removed some dead code from Myre (version bump to 10.0.0 since some of the dead code was in the public API surface) Updated packages Added AnimatedDude2.cs to test the new blocky dude

+++ Myre/Myre/Extensions/GameTimeExtensions.cs
@@ -21,7 +21,7 @@ namespace Myre.Extensions
         /// </summary>
         /// <param name="t">The t.</param>
         /// <returns>The total number of seconds elapsed since the game started.</returns>
-        public static float TotalSeconds(this Microsoft.Xna.Framework.GameTime t)
+        public static double TotalSeconds(this Microsoft.Xna.Framework.GameTime t)
         {
             return (float)t.TotalGameTime.TotalSeconds;
         }
